Updated Survey API

+++ MiniSurvey.Client/Dto/OpinionPoll.cs
@@ -7,7 +7,7 @@ namespace MiniSurvey.Client.Dto
 {
     public class OpinionPoll
     {
-        public int OptionId { get; set; }
+        public long OptionId { get; set; }
         public string Option { get; set; }
         public double Percentage { get; set; }
     }
