+++ MiniSurvey.Client/Models/QuestionOption.cs
@@ -8,8 +8,8 @@ namespace MiniSurvey.Client.Models
     public partial class QuestionOption
     {
         public int Id { get; set; }
-        public int QuestionId { get; set; }
-        public int OptionId { get; set; }
+        public long QuestionId { get; set; }
+        public long OptionId { get; set; }
 
         public virtual Option Option { get; set; }
         public virtual Question Question { get; set; }
